% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructCTnet.R
\name{ConstructCTnet}
\alias{ConstructCTnet}
\title{ConstructCTnet for construct cell-specific PPI}
\usage{
ConstructCTnet(
  species = "mmu",
  min.pct = 0.1,
  intersect = 0.05,
  lib = NULL,
  myexp = NULL,
  ctcell = NULL,
  log2fc = 0.1,
  imput = TRUE,
  graphdata = NULL,
  seedgene = NULL,
  mysigDEG = NULL,
  celltype = NULL,
  k = 15,
  thread_no = 2,
  maxiter = 1000,
  reso1 = 0.6,
  valuek = 50,
  valuen = 100
)
}
\arguments{
\item{species}{character,define the species, should be 'mmu' or 'human'}

\item{min.pct}{numeric, default is 0.1 (10\%). Percentage of gene expression: genes with an expression percentage below this value will be filtered out}

\item{intersect}{numeric, default is 0.05 (5\%). Percentage of gene co-expression: gene pairs with a co-expression percentage below this value will be filtered out}

\item{lib}{character, default is NULL. The lib must be NULL (Drugbank database) or herb (HERB database)}

\item{myexp}{gene expression matrix (dgMatrix) extracted from SeuratObject}

\item{ctcell}{character vector, cells within one cell-type}

\item{log2fc}{numeric, default is 0.1. Genes whose |log2fc| greater than this threshold are considered as differentially expressed genes}

\item{imput}{logical, whether compute imputation. If imputation has been completed and you want to change log2fc or other parameters and rerun this function, define the imput=FALSE and give your graphdata.rda filepath to graphdata.}

\item{graphdata}{character vector, your graphdata.rda filepath. If imput=FALSE, you must give your graphdata.rda filepath to graphdata}

\item{seedgene}{character vector, if seedgene is NULL, we use differentially expressed psychiatry diseases genes as seedgene, you can also provide a geneset as seedgene}

\item{mysigDEG}{dataframe, a dataframe given your differentially expressed genes which must has column: gene,NewType,avg_log2FC in which NewType column defining your celltypes}

\item{celltype}{character, define celltype, must be consistent with the celltype you provided to mysigDEG}

\item{k}{numeric, default is 15.  Number of nearest neighbors used in the ACTIONet decomposition. for improving computational speed we use one k without performing a search over a range of k. the larger the value of k, the slower the computation}

\item{thread_no}{numeric, default is 2. Number of CPU threads to use for parallel computation}

\item{maxiter}{numeric, default is 1000. Maximum number of iterations allowed for the random walk with restart (RWR) algorithm. The iteration stops earlier if convergence is reached, as determined by the epsilon threshold.}

\item{reso1}{numeric, default is 0.6. Resolution of louvain algorithm identifying modules}

\item{valuek}{numeric, default is 50. The expected number of genes drawn as seed genes each time}

\item{valuen}{numeric, default is 100. Total number of genes drawn each time}
}
\value{
CTNet Class
}
\description{
ConstructCTnet for construct cell-specific PPI
}
