% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructCTnet.R
\docType{class}
\name{CTNet-class}
\alias{CTNet-class}
\alias{show,CTNet-method}
\title{CTNet Class}
\usage{
\S4method{show}{CTNet}(object)
}
\arguments{
\item{object}{CTNet class}
}
\description{
S4 class for storing cell-type network.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(CTNet)}: show method

}}
\section{Slots}{

\describe{
\item{\code{features}}{A list of features.}

\item{\code{ctnets}}{A list of cell-type-specific networks.}

\item{\code{ctmodule}}{A list of modules.}

\item{\code{RWR}}{A list of random walk results.}
}}

