% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRPS.R
\name{cutproximity_herb}
\alias{cutproximity_herb}
\title{proximity process}
\usage{
cutproximity_herb(
  proximity = NULL,
  zcutoff = NULL,
  ctnet = NULL,
  filename = NULL
)
}
\arguments{
\item{proximity}{dataframe}

\item{zcutoff}{numeric}

\item{ctnet}{CTNet}

\item{filename}{character}
}
\description{
Specify your cutoff of proximity value computed by herb, not export
}
