% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizition.R
\name{DB2Cyto}
\alias{DB2Cyto}
\title{output for cytoscape}
\usage{
DB2Cyto(drugs = NULL, proximity = NULL, ctnets = NULL, hop = 3)
}
\arguments{
\item{drugs}{character vector, give the drugs you want to output}

\item{proximity}{dataframe, the proximity file yielding by DrugEft function}

\item{ctnets}{ctnet CTNet Class, yielding by ConstructCTnet}

\item{hop}{numeric, default is 3. Maximum steps considered for computing network proximity.}
}
\description{
output for cytoscape
}
