% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRPS.R
\name{DrugEft}
\alias{DrugEft}
\title{Network Proximity computation}
\usage{
DrugEft(
  species = "mmu",
  s = 0.2,
  log2fc = 0.1,
  lib = NULL,
  ctnet = NULL,
  celltype = NULL,
  prox = TRUE,
  myprox = NULL,
  mysigDEG = mysigDEG,
  hop = 3,
  zcutoff = -0.5,
  sankyn = 20,
  sankycols = NULL
)
}
\arguments{
\item{species}{character,define the species, should be 'mmu' or 'human'}

\item{s}{numeric, default is 0.2. Shrinkage coefficient, ranging from 0 to 1. The smaller the value, the smaller the difference in weights between different scoring levels.}

\item{log2fc}{numeric, default is 0.1. Genes whose |log2fc| greater than this threshold are considered as differentially expressed genes}

\item{lib}{character, default is NULL. The lib must be NULL (Drugbank database) or herb (HERB database)}

\item{ctnet}{CTNet Class, output by ConstructCTnet}

\item{celltype}{character, define celltype, must be consistent with the celltype you provided to mysigDEG}

\item{prox}{logical, whether computing proximity, once you have run this function, you will get the raw proximity value without cutoff (named yourcelltype.proximity_raw.rda). If you want to change the cutoff but do not compute proximity value again for reducing computational time, you can just define prox=FALSE and give the proximity_raw.rda filepath to myprox.}

\item{myprox}{character, yourcelltype.proximity_raw.rda filepath}

\item{mysigDEG}{dataframe, a dataframe given your differentially expressed genes which must has column: gene,NewType,avg_log2FC in which NewType column defining your celltypes}

\item{hop}{numeric, default is 3. Maximum steps considered for computing network proximity.}

\item{zcutoff}{numeric, default is 0.5. Cutoff of proximity value to compute Drug repurposing score.}

\item{sankyn}{numeric, drug number for sankyplot}

\item{sankycols}{character vector, default is NULL for using the colors we provide. You can provide your color codes to sankycols}
}
\value{
data.frame
}
\description{
Network Proximity computation
}
