% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizition.R
\name{keygrep}
\alias{keygrep}
\title{extract drugs}
\usage{
keygrep(key = NULL, keytype = "Description", celltype = NULL, cutoff)
}
\arguments{
\item{key}{character, the phenotype you want to grep}

\item{keytype}{character, grep key of column: Description or Details}

\item{celltype}{character}

\item{cutoff}{numeric}
}
\description{
extract drugs
}
