% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizition.R
\name{netVisual}
\alias{netVisual}
\title{plot NetworkModule PPI}
\usage{
netVisual(
  ctnet = NULL,
  filename = NULL,
  SamType = NULL,
  labelgene = NULL,
  cols = NULL,
  width = 5,
  height = 5
)
}
\arguments{
\item{ctnet}{CTNet Class, output by ConstructCTnet}

\item{filename}{character, define your filename by your celltype}

\item{SamType}{character, define your filename by your sample phenotype}

\item{labelgene}{character, default is NULL with the seed/target genes ploted. You can provide the label genes you want to plot}

\item{cols}{character, default is NULL with the colors we provide. You can provide colors as you like}

\item{width}{output pdf width}

\item{height}{output pdf height}
}
\value{
none
}
\description{
plot NetworkModule PPI
}
