% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizition.R
\name{plotsanky}
\alias{plotsanky}
\title{plot your sanky to visual drug-target-module}
\usage{
plotsanky(
  dfdrug,
  ctnets,
  filename = "Celltype",
  pattle = NULL,
  width = 6,
  height = 10
)
}
\arguments{
\item{dfdrug}{dataframe, dataframe yielding by DrugEft containing drugs you want to plot}

\item{ctnets}{ctnet CTNet Class, output by ConstructCTnet}

\item{filename}{character, define your filename by your celltype}

\item{pattle}{character vector, default is NULL with the colors we provide. You can also give your colors pattle.}

\item{width}{output pdf width}

\item{height}{output pdf height}
}
\value{
none
}
\description{
plot your sanky to visual drug-target-module
}
