% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizition.R
\name{Targetppi}
\alias{Targetppi}
\title{Targetppi, not export}
\usage{
Targetppi(module = NULL, targets = NULL, hop = hop, ctnet = ctnets)
}
\arguments{
\item{module}{character, module}

\item{targets}{character vector, target genes}

\item{hop}{numeric, hop step}

\item{ctnet}{CTNet Class}
}
\description{
Targetppi, not export
}
